+++ alonet/deformable_detr/ops/test.py
@@ -25,8 +25,8 @@ load_MultiScaleDeformableAttention()
 
 N, M, D = 1, 2, 2
 Lq, L, P = 2, 2, 2
-shapes = torch.as_tensor([(6, 4), (3, 2)], dtype=torch.long).cuda()
-level_start_index = torch.cat((shapes.new_zeros((1,)), shapes.prod(1).cumsum(0)[:-1]))
+shapes = torch.as_tensor([(6, 4), (3, 2)], dtype=torch.int32).cuda()
+level_start_index = torch.cat((shapes.new_zeros((1,)), shapes.prod(1).cumsum(0)[:-1])).type(torch.int32)
 S = sum([(H * W).item() for H, W in shapes])
